Added a new hosted service that uses the DigitalstromClient library and its SceneClient class to start threads listening for digitalstrom events.
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C420

Added DigitalstromConnectionProvider interface for encapsulating all necessary info for connecting to a digitalstrom dss server, including the uris, auth info and optionally the proxy to use. It now also allows for skipping the authentication via dss token mechanism altogether if the auth is already included in the uri (like with the ds-tools.net/cloudredir.php proxy).

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C421

Implemented a ConnectionProvider for digitalstrom client library and a subclass of DigitalstromAuth that uses the TokenStore internally. Added this conn provider as singleton service for IoC. Removed connection infos from DigitalstromEventsHostedService and added parameter for dependency injection.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C422

Simplfied SceneClient constructor and allow for direct passing of a connection provider.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C423