Implemented coarse stub of event handler in DssSceneCaller
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C540

Added internal event and setter for avoiding circular events (when a dss event triggers a scene call, which triggers a dss event etc.)

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C541

Completed unit test for changing the Twin and waiting for SceneCall api calls. Added unit test helper methods for mocking SceneCall events and waiting for the SceneCall to happen. Initialize RoomState with empty dictionaries to give subscribers a chance to catch all Add events and subscribe to the elements further down the tree.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C542

Implemented subscribing of events to already preexisting model tree elements. Added unit test for this case.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C543

Added unit test for DigitalstromTwin (not working yet), with tests for both one-ways and the two-way sync.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C544