Sync dbcontext access in case of parallel thread usage of the same token store.
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C580

Avoid creating a scope twice and also nested into each other

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C581

Warning: this commit does not build! It is used to save the intermediate work and restore it on another computer.

Initial check in of integration test for DigitalstromHost including a web host factory that facilitates an in memory database for creating a db context to work in the tests.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C582