Do not subscribe to events double (duplicates in eventsToSubscribe), also wait for initialization of subscriber before starting with polling.
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C601

Moved duplicate model syncing code out of DssEventSubscriber (which is now only the subscription thread and action worker thread owner) and into DigitalstromDssTwin which is now responsible for synching the model. Unit tests do not yet work - I suggest the tests are right and the bug is still in the code not the tests, but I got no time to track this down, so this broken version is checked in for having it in the history.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C602

Made AbstractState value reading and writing threadsafe, as it gets written and read concurrently by the action worker thread in case of the Twin.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C603