Fixed locking problem with DssEventSubscriber - the semaphore was only released once and then locked forever *facepalm*. Some timing improvements for stability in unit tests and some minor code cleanup.
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C604

Removed worker thread for producer/consumer queueing of api client calls from WebApiWorkerThreadBase, as it is not really needed. Renamed it into LongPollingClientBase as this is the only responsibility that is encapsuled in this class from now on. Adjusted all uses of the new class accordingly.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C605

corrected Dispose method of Twin class.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C606

Added property tree query methods for retrieving the output channel types/ids for single devices as well as their last cached state in the dss. Added classes for strongly typed deserializing of these query responses.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C607

Added unit tests for the two new query methods in dssclient.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C608