Changed resampler internal time series from TimeSeriesStream to normal TimeSeries due to overflow issues when aggregating for a whole day.
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C654

Found bug and wrote a unit test reproducing it - when the resampling rate is so high that multiple time series will be aggregated into a single bucket of the result time series, the aggregation fails, as all series are applied in series after each other.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C655

Created a first stable non-optimized version of the resampler that fixes the bug with multiple time series for a single resampled bucket.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C656

Optimized form (breaks if the time bucket pointer moved past the resampled bucket) of last improvement of TimeSeriesResampler.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C657

Implemented a two-step reducer and resampler for energy values, such that we only have to read the db once, only decompress all data once, do not have to store the full value list in RAM at any time, but still have the bug fixed where the reduction factor spans across multiple days. Also, the GraphCollectionViewModelBase now uses the IEnumerable<TimeSeries> method of the resampler per default. Adjusted other ViewModels as well where needed.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C658

Removed unused/obsolete Resampler methods.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C659

Built in a force overwrite for maxDataPoints as well as changed all TimeSeriesStream to TimeSeries in Sonnen ViewModel due to overflow issues when aggregating.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C660