Added db context for netatmo host and added classes for the general list of devices, modules and measures as classes for storing the actual data both midres (5 min, which is the maximum resolution of the netatmo devices) and lowres (160 min interval per month).
git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C753

Removed temporary block for research purposes and implemented results in NetatmoModuleMeasure, where the measure type can be set and the decimal precision is set accordingly.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C754

Added a structure service singleton for netatmo devices and modules and simplified code in polling service accordingly.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C755

Added api to trigger polling of netatmo values manually.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C756

Implemented saving to database of mid resolution netatmo values. Added method to query for a module and device name in structure service.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C757

Added netatmo db sets to persistence context and added migration. Fixed a few small bugs, such as name getters in structure service and missing saveDb call.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C758

Added a mapping of netatmo modules and measures to database guids in structure service.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C759

Made api more robust: only output non-empty distinct values.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C760

Added netatmo viewmodel.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C761

Added method in polling service that is much better in mapping a non-equidistant time series that is returned by the netatmo api to the equidistant one that we are storing than the naive appoach that was used before. It checks if the preferred bucket is already filled and tries to either move to a previous one that is still empty or push it to the next one, depending on the exact timestamp and the filling of prev and next buckets.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C762

Added exception when reaching the api rate limit.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C763

Implemented low resolution resampling and saving in netatmo polling service. small code cleanup by extracting method for polling values from netatmo server api.

git-tfs-id: [https://philipdaubmeier.visualstudio.com]$/SmarthomeAPI;C764